+++ src/main/kotlin/net/tarasandedevelopment/tarasande/system/feature/modulesystem/impl/render/ModuleBedESP.kt
@@ -204,7 +204,7 @@ class ModuleBedESP : Module("Bed ESP", "Highlights all beds", ModuleCategory.REN
 
     object Breaker {
 
-        private val breakSpeed = BiFunction<Node, Node, Double> { _, movement -> PlayerUtil.getBreakSpeed(BlockPos(movement.x, movement.y, movement.z)).first }
+        private val breakSpeed = BiFunction<Node, Node, Double> { _, movement -> PlayerUtil.getBreakSpeed(BlockPos(movement.x, movement.y, movement.z)).first.toDouble() }
         private var defenders: List<BlockPos>? = null
 
         private val pathFinder = PathFinder({ _, node -> defenders?.contains(BlockPos(node.x, node.y, node.z)) == true }, cost = breakSpeed)
